Calendar and AddEvent Navigation and form
Fixed navigation for Calendar and AddEvent Pages, created simple (nonfunctional) form to add new events
