+++ Views/Event/Calendar.cshtml
@@ -1,0 +1,21 @@
+﻿@model Event
+@{
+    Layout = null;
+}
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta name="viewport" content="width=device-width" />
+        <title> Calendar</title>
+    </head>
+<body>
+    <h1>Calendar</h1>
+</body>
+<div class="row">
+    <div class="col-sm-4">
+        <div class="list-group">
+            <a asp-controller="Event" asp-action="AddEvent"
+               class="list-group-item">Add New Event</a>
+        </div>
+    </div>
+</html>
