--- Views/Home/AddEvent.cshtml
@@ -1,37 +1,0 @@
-﻿@model EventViewModel
-@{
-    Layout = null;
-}
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta name="viewport" content="width-device-width" />
-        <title>Add New Event</title>
-        <style>
-            /* add CSS styles here */
-        </style>
-    </head>
-    <body>
-        <h1>Add New Event</h1>
-        <form asp-action="AddEvent" method="post">
-            <div>
-                <label asp-for="EventLabel">Event Label</label>
-                <input asp-for="EventLabel" />
-            </div>
-            <div>
-                <label asp-for="EventDescription">Event Description</label>
-                <input asp-for="EventDescription" />
-            </div>
-            <div>
-                <label asp-for="EventLength">Event Length</label>
-                <input asp-for="EventLength" />
-            </div>
-        
-            <div>
-                <button type="submit">AddEvent</button>
-                <a asp-action="AddEvent">Clear</a>
-            </div>
-        </form>
-    </body>
-</html>
-
