--- Views/Home/Calendar.cshtml
@@ -1,7 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Calendar";
-}
-
-<div>
-    <img src="~/images/CalendarPlaceHolder.png" alt="Calendar Place Holder">
-</div>
