--- Views/Home/Dashboard.cshtml
@@ -1,7 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Dashboard";
-}
-
-<div>
-        <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Profile"><img src="~/images/DashboardPlaceHolder.png" alt="Dashboard Place Holder"></a>
-</div>
