+++ Views/Home/Index.cshtml
@@ -1,5 +1,6 @@
 ﻿@{
-    Layout = null;
+    ViewData["Title"] = "Dashboard";
+
 }
 <!DOCTYPE html>
 <html>
